feat: switch PhraseId from 64-bit to 128-bit FNV-1a hash
128-bit hashes make collisions effectively impossible for any realistic
number of phrases (~26 billion phrases needed for 1-in-a-billion collision
probability vs ~190 for 64-bit). The tradeoff is 16 bytes per PhraseId
instead of 8, which is negligible for the use case.

Changes:
- PhraseId inner type: u64 -> u128
- Hash function: fnv1a_hash_str_64 -> fnv1a_hash_str_128
- Renamed as_u64() to as_u128()
- Updated all id-based lookup signatures across registry, locale, and error types
- Display format widened to 32 hex chars
- Updated documentation in DESIGN.md and APPENDIX_RUST_INTEGRATION.md

+++ crates/rlf/src/interpreter/error.rs
@@ -169,7 +169,7 @@ pub enum EvalError {
 
     /// Phrase not found by PhraseId hash.
     #[error("phrase not found for id: {id}")]
-    PhraseNotFoundById { id: u64 },
+    PhraseNotFoundById { id: u128 },
 
     /// Required variant key is missing from phrase.
     #[error("missing variant '{key}' in phrase '{phrase}', available: {}{}", available.join(", "), format_suggestions(suggestions))]
